requirements.txt
pandas==2.2.3
typing==3.7.4.3
openai==1.98.0
SQLAlchemy==2.0.42
python-dotenv==1.1.0
smolagents==1.20.0
pydantic==2.11.7
json_repair==0.48.0